cleanup and move test_ota_client_stub as test_servicer
